Try to make match-all non-quadratic
The approach implemented here does not work for the following
reason: matches are delayed by one character to account for
simple lookahead assertions (like $ for example). This causes
states to be added to the visited set at the end position of the
match, which will be the starting position of the next search.
Because of this, the assumption that the new search is not messed
up by the preceeding search is invalid.

One way to fix this might be to clear all state bits at the
starting position of the input instead of the entire bitvector.
This would remove quadraticity in the haystack length, but it
would still be quadratic in the regex length, so it is questionable
whether it's worth going through the hassle of implementing it.